pyproject.toml
[tool.poetry]
name = "wifienable"
version = "0.1.0"
description = ""
authors = ["Aucke Bos <aucke.bos97@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
selenium = "^4.8.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
